move amp middle files (#977)
