+++ src/views/amp/index.amp.ejs
@@ -10,7 +10,7 @@
     <script async custom-element="amp-iframe" src="https://cdn.ampproject.org/v0/amp-iframe-0.1.js"></script>
     <script async custom-element="amp-ad" src="https://cdn.ampproject.org/v0/amp-ad-0.1.js"></script>
     <script async custom-element="amp-analytics" src="https://cdn.ampproject.org/v0/amp-analytics-0.1.js"></script>
-    <title><%= storyInfo.storyTitle %> - 鏡週刊</title>
+    <title><%= storyInfo.storyTitle %></title>
     <link rel="canonical" href="<%= storyInfo.storyURL %>">
     <meta property="og:title" content="<%= storyInfo.storyTitle %> - 鏡週刊" >
     <meta property="og:url" content="<%= storyInfo.storyURLAMP %>">
